voice_recog/requirements.txt
asgiref==3.6.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
google-api-core==2.11.0
google-auth==2.16.2
google-cloud-speech==2.18.0
googleapis-common-protos==1.59.0
grpcio==1.51.3
grpcio-status==1.51.3
idna==3.4
numpy==1.24.1
Pillow==9.4.0
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
PyOpenGL @ file:///C:/Users/ACPC/Downloads/PyOpenGL-3.1.6-cp39-cp39-win_amd64.whl
pypng==0.20220715.0
pytz==2022.7.1
requests==2.28.2
rsa==4.9
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
Wave==0.0.2
